requirements.txt
redis~=7.0.1
mysql-connector-python